Skip auth when posting to stop_impersonating
Auth checks if the current_ability using the impersonated current_user `can? :manage, User` causing an AccessDenied error blocking the user from ending the impersonation.  stop_impersonating should be safe to call as any user so just skip the auth check.  Alternatively a `true_ability` could be setup similar to `true_user` which could be used for the `can?` check in `:auth`